+++ src-docs/src/views/range/dual_range.js
@@ -1,12 +1,9 @@
 import React, {
   Component,
-  Fragment,
 } from 'react';
 
 import {
   EuiDualRange,
-  EuiSpacer,
-  EuiFormHelpText,
 } from '../../../../src/components';
 
 import makeId from '../../../../src/components/form/form_row/make_id';
@@ -15,21 +12,8 @@ export default class extends Component {
   constructor(props) {
     super(props);
 
-    this.levels = [
-      {
-        min: 0,
-        max: 600,
-        color: 'danger'
-      },
-      {
-        min: 600,
-        max: 2000,
-        color: 'success'
-      }
-    ];
-
     this.state = {
-      value: [120, 480]
+      value: ['', ''],
     };
   }
 
@@ -41,96 +25,15 @@ export default class extends Component {
 
   render() {
     return (
-      <Fragment>
-
-        <EuiDualRange
-          id={makeId()}
-          min={0}
-          max={2000}
-          value={this.state.value}
-          onChange={this.onChange}
-          aria-label="Use aria labels when no actual label is in use"
-          showLabels
-          name="dualRange"
-        />
-
-        <EuiSpacer size="xl" />
-
-        <EuiDualRange
-          id={makeId()}
-          min={0}
-          max={2000}
-          value={this.state.value}
-          onChange={this.onChange}
-          disabled
-          aria-label="Use aria labels when no actual label is in use"
-          showLabels
-        />
-
-        <EuiSpacer size="xl" />
-
-        <EuiDualRange
-          id={makeId()}
-          min={0}
-          max={2000}
-          value={this.state.value}
-          onChange={this.onChange}
-          aria-label="Use aria labels when no actual label is in use"
-          showLabels
-          showInput
-          showRange
-        />
-
-        <EuiSpacer size="xl" />
-
-        <EuiDualRange
-          id={makeId()}
-          min={0}
-          max={2000}
-          step={50}
-          value={this.state.value}
-          onChange={this.onChange}
-          aria-label="Use aria labels when no actual label is in use"
-          aria-describedby="levelsHelp"
-          showLabels
-          showInput
-          compressed
-          levels={this.levels}
-        />
-        <EuiFormHelpText id="levelsHelp">Recommended levels are 600 and above.</EuiFormHelpText>
-
-        <EuiSpacer size="xl" />
-
-        <EuiDualRange
-          id={makeId()}
-          min={0}
-          max={2000}
-          step={50}
-          value={this.state.value}
-          onChange={this.onChange}
-          aria-label="Use aria labels when no actual label is in use"
-          showTicks
-          showRange
-          tickInterval={300}
-        />
-
-        <EuiSpacer size="xl" />
-
-        <EuiDualRange
-          id={makeId()}
-          min={0}
-          max={2000}
-          step={50}
-          value={this.state.value}
-          onChange={this.onChange}
-          aria-label="Use aria labels when no actual label is in use"
-          aria-describedby="levelsHelp"
-          showTicks
-          showInput
-          tickInterval={500}
-          levels={this.levels}
-        />
-      </Fragment>
+      <EuiDualRange
+        id={makeId()}
+        min={-100}
+        max={200}
+        step={10}
+        value={this.state.value}
+        onChange={this.onChange}
+        showLabels
+      />
     );
   }
 }