+++ src/components/badge/notification_badge/badge_notification.tsx
@@ -1,19 +1,40 @@
-import React, { HTMLAttributes, ReactNode, SFC } from 'react';
+import React, { HTMLAttributes, ReactNode, FunctionComponent } from 'react';
 import classNames from 'classnames';
-import { CommonProps } from '../../common';
+import { CommonProps, keysOf } from '../../common';
+
+const colorToClassMap: { [color: string]: string | null } = {
+  accent: null,
+  subdued: 'euiNotificationBadge--subdued',
+};
+
+export const COLORS: BadgeNotificationColor[] = keysOf(colorToClassMap);
+export type BadgeNotificationColor = keyof typeof colorToClassMap;
+
+const sizeToClassNameMap = {
+  s: null,
+  m: 'euiNotificationBadge--medium',
+};
+
+export const SIZES: BadgeNotificationSize[] = keysOf(sizeToClassNameMap);
+export type BadgeNotificationSize = keyof typeof sizeToClassNameMap;
 
 export interface EuiNotificationBadgeProps
   extends CommonProps,
     HTMLAttributes<HTMLSpanElement> {
-  children?: ReactNode;
+  children: ReactNode;
+  size?: BadgeNotificationSize;
+  color?: BadgeNotificationColor;
 }
 
-export const EuiNotificationBadge: SFC<EuiNotificationBadgeProps> = ({
-  children,
-  className,
-  ...rest
-}) => {
-  const classes = classNames('euiNotificationBadge', className);
+export const EuiNotificationBadge: FunctionComponent<
+  EuiNotificationBadgeProps
+> = ({ children, className, size = 's', color = 'accent', ...rest }) => {
+  const classes = classNames(
+    'euiNotificationBadge',
+    sizeToClassNameMap[size],
+    colorToClassMap[color],
+    className
+  );
 
   return (
     <span className={classes} {...rest}>