+++ src/components/color_picker/color_picker_swatch.tsx
@@ -1,4 +1,4 @@
-import React, { HTMLAttributes, SFC } from 'react';
+import React, { HTMLAttributes, FunctionComponent } from 'react';
 import classNames from 'classnames';
 import { EuiColorPickerEmptySwatch } from './color_picker_empty_swatch';
 
@@ -7,11 +7,9 @@ export interface EuiColorPickerSwatchProps
   color?: string;
 }
 
-export const EuiColorPickerSwatch: SFC<EuiColorPickerSwatchProps> = ({
-  color,
-  className,
-  ...rest
-}) => {
+export const EuiColorPickerSwatch: FunctionComponent<
+  EuiColorPickerSwatchProps
+> = ({ color, className, ...rest }) => {
   const isClear = !color;
   const classes = classNames('euiColorPicker__swatch', className, {
     euiColorPicker__emptySwatch: isClear,