+++ src/components/form/form_row/index.d.ts
@@ -1,6 +1,6 @@
 import { CommonProps } from '../../common';
 
-import { SFC, ReactNode, HTMLAttributes } from 'react';
+import { FunctionComponent, ReactNode, HTMLAttributes } from 'react';
 
 declare module '@elastic/eui' {
   /**
@@ -20,5 +20,5 @@ declare module '@elastic/eui' {
       displayOnly?: boolean;
     };
 
-  export const EuiFormRow: SFC<EuiFormRowProps>;
+  export const EuiFormRow: FunctionComponent<EuiFormRowProps>;
 }