+++ src/components/form/range/range_track.js
@@ -10,24 +10,55 @@ import { EuiRangeTicks } from './range_ticks';
 export { LEVEL_COLORS };
 
 export class EuiRangeTrack extends Component {
+  validateValueIsInStep = (value) => {
+    if (value < this.props.min) {
+      throw new Error(`The value of ${value} is lower than the min value of ${this.props.min}.`);
+    }
+    if (value > this.props.max) {
+      throw new Error(`The value of ${value} is higher than the max value of ${this.props.max}.`);
+    }
+    // Error out if the value doesn't line up with the sequence of steps
+    if ((value - this.props.min) % this.props.step > 0) {
+      throw new Error(`The value of ${value} is not included in the possible sequence provided by the step of ${this.props.step}.`);
+    }
+    // Return the value if nothing fails
+    return value;
+  }
 
-  calculateTicksObject = (min, max, interval) => {
-    // Calculate the width of each tick mark
-    const tickWidthDecimal = (interval / ((max - min) + interval));
-    const tickWidthPercentage = tickWidthDecimal * 100;
 
-    // Loop from min to max, creating ticks at each interval
+  calculateSequence = (min, max, interval) => {
+    // Loop from min to max, creating adding values at each interval
     // (adds a very small number to the max since `range` is not inclusive of the max value)
     const toBeInclusive = .000000001;
-    const sequence = range(min, max + toBeInclusive, interval);
+    return range(min, max + toBeInclusive, interval);
+  }
 
-    return (
-      {
-        decimalWidth: tickWidthDecimal,
-        percentageWidth: tickWidthPercentage,
-        sequence: sequence,
-      }
-    );
+  calculateTicks = (min, max, step, tickInterval, customTicks) => {
+    let ticks;
+
+    if (customTicks) {
+      // If custom values were passed, use those for the sequence
+      // But make sure they align with the possible sequence
+      ticks = customTicks.map(tick => {
+        return this.validateValueIsInStep(tick.value);
+      });
+    } else {
+      // If a custom interval was passed, use those for the sequence
+      // But make sure they align with the possible sequence
+      const interval = tickInterval || step;
+      const tickSequence = this.calculateSequence(min, max, interval);
+
+      ticks = tickSequence.map(tick => {
+        return this.validateValueIsInStep(tick);
+      });
+    }
+
+    // Error out if there are too many ticks to render
+    if (ticks.length > 20) {
+      throw new Error(`The number of ticks to render is too high (${ticks.length}), reduce the interval.`);
+    }
+
+    return ticks;
   }
 
   render() {
@@ -39,22 +70,25 @@ export class EuiRangeTrack extends Component {
       step,
       showTicks,
       tickInterval,
-      ticks, // eslint-disable-line no-unused-vars
+      ticks,
       levels,
       onChange,
       value
     } = this.props;
 
-    let tickObject;
+    // TODO: Move these to only re-calculate if no-value props have changed
+    this.validateValueIsInStep(max);
+
+    let tickSequence;
     const inputWrapperStyle = {};
     if (showTicks) {
-      tickObject = this.calculateTicksObject(min, max, tickInterval || step || 1);
+      tickSequence = this.calculateTicks(min, max, step, tickInterval, ticks);
 
       // Calculate if any extra margin should be added to the inputWrapper
       // because of longer tick labels on the ends
-      const lengthOfMinLabel = String(tickObject.sequence[0]).length;
-      const lenghtOfMaxLabel = String(tickObject.sequence[tickObject.sequence.length - 1]).length;
-      const isLastTickTheMax = tickObject.sequence[tickObject.sequence.length - 1] === max;
+      const lengthOfMinLabel = String(tickSequence[0]).length;
+      const lenghtOfMaxLabel = String(tickSequence[tickSequence.length - 1]).length;
+      const isLastTickTheMax = tickSequence[tickSequence.length - 1] === max;
       if (lengthOfMinLabel > 2) {
         inputWrapperStyle.marginLeft = `${(lengthOfMinLabel / 5)}em`;
       }
@@ -70,7 +104,7 @@ export class EuiRangeTrack extends Component {
     return (
       <div className={trackClasses} style={inputWrapperStyle}>
         {children}
-        {!!levels.length && (
+        {levels && !!levels.length && (
           <EuiRangeLevels
             levels={levels}
             max={max}
@@ -78,14 +112,16 @@ export class EuiRangeTrack extends Component {
             showTicks={showTicks}
           />
         )}
-        {showTicks && (
+        {tickSequence && (
           <EuiRangeTicks
             disabled={disabled}
             onChange={onChange}
             ticks={ticks}
-            tickObject={tickObject}
+            tickSequence={tickSequence}
             value={value}
+            min={min}
             max={max}
+            interval={tickInterval || step}
           />
         )}
       </div>