+++ src/components/table/index.d.ts
@@ -4,12 +4,13 @@ import { IconType } from '../icon';
 import { HorizontalAlignment } from '../../services/alignment';
 
 import {
-  SFC,
+  FunctionComponent,
   HTMLAttributes,
   TableHTMLAttributes,
   ButtonHTMLAttributes,
   ThHTMLAttributes,
-  TdHTMLAttributes
+  TdHTMLAttributes,
+  ReactNode
 } from 'react';
 
 declare module '@elastic/eui' {
@@ -24,7 +25,7 @@ declare module '@elastic/eui' {
     responsive?: boolean;
   }
 
-  export const EuiTable: SFC<
+  export const EuiTable: FunctionComponent<
     CommonProps & TableHTMLAttributes<HTMLTableElement> & EuiTableProps
   >;
 
@@ -36,7 +37,7 @@ declare module '@elastic/eui' {
 
   export interface EuiTableBodyProps {}
 
-  export const EuiTableBody: SFC<CommonProps & EuiTableBodyProps>;
+  export const EuiTableBody: FunctionComponent<CommonProps & EuiTableBodyProps>;
 
   /**
    * table header type defs
@@ -46,7 +47,7 @@ declare module '@elastic/eui' {
 
   export interface EuiTableHeaderProps {}
 
-  export const EuiTableHeader: SFC<CommonProps & EuiTableHeaderProps>;
+  export const EuiTableHeader: FunctionComponent<CommonProps & EuiTableHeaderProps>;
 
   /**
    * table header button type defs
@@ -58,7 +59,7 @@ declare module '@elastic/eui' {
     iconType?: IconType;
   }
 
-  export const EuiTableHeaderButton: SFC<
+  export const EuiTableHeaderButton: FunctionComponent<
     CommonProps &
       ButtonHTMLAttributes<HTMLButtonElement> &
       EuiTableHeaderButtonProps
@@ -79,9 +80,15 @@ declare module '@elastic/eui' {
     isSorted?: boolean;
     isSortAscending?: boolean;
     scope?: TableHeaderCellScope;
+    isMobileHeader?: boolean;
+    hideForMobile?: boolean;
+    mobileOptions?: {
+      show?: boolean;
+      only?: boolean;
+    };
   }
 
-  export const EuiTableHeaderCell: SFC<
+  export const EuiTableHeaderCell: FunctionComponent<
     CommonProps &
       ThHTMLAttributes<HTMLTableHeaderCellElement> &
       EuiTableHeaderCellProps
@@ -104,7 +111,7 @@ declare module '@elastic/eui' {
     scope?: EuiTableHeaderCellCheckboxScope;
   }
 
-  export const EuiTableHeaderCellCheckbox: SFC<
+  export const EuiTableHeaderCellCheckbox: FunctionComponent<
     CommonProps &
       TdHTMLAttributes<HTMLTableCellElement> &
       EuiTableHeaderCellCheckboxProps
@@ -120,7 +127,7 @@ declare module '@elastic/eui' {
     isSelected?: boolean;
   }
 
-  export const EuiTableRow: SFC<
+  export const EuiTableRow: FunctionComponent<
     CommonProps & EuiTableRowProps & HTMLAttributes<HTMLTableRowElement>
   >;
 
@@ -130,21 +137,34 @@ declare module '@elastic/eui' {
    * @see './table_row_cell.js'
    */
 
-  export interface EuiTableRowCellProps {
+  interface EuiTableRowCellSharedPropsShape {
     align?: HorizontalAlignment;
+    showOnHover?: boolean;
+    textOnly?: boolean;
+    truncateText?: boolean;
+  }
+
+  export type EuiTableRowCellMobileOptionsShape = {
+    show?: boolean;
+    only?: boolean;
+    render?: ReactNode;
+    header?: ReactNode | boolean;
+    enlarge?: boolean;
+    fullWidth?: boolean;
+  }
+
+  export interface EuiTableRowCellProps {
     hasActions?: boolean;
     header?: string;
     hideForMobile?: boolean;
     isExpander?: boolean;
     isMobileFullWidth?: boolean;
     isMobileHeader?: boolean;
-    showOnHover?: boolean;
-    textOnly?: boolean;
-    truncateText?: boolean;
+    mobileOptions?: EuiTableRowCellMobileOptionsShape & EuiTableRowCellSharedPropsShape;
   }
 
-  export const EuiTableRowCell: SFC<
-    CommonProps & TdHTMLAttributes<HTMLTableCellElement> & EuiTableRowCellProps
+  export const EuiTableRowCell: FunctionComponent<
+    CommonProps & TdHTMLAttributes<HTMLTableCellElement> & EuiTableRowCellSharedPropsShape & EuiTableRowCellProps
   >;
 
   /**
@@ -155,7 +175,7 @@ declare module '@elastic/eui' {
 
   export interface EuiTableRowCellCheckboxProps {}
 
-  export const EuiTableRowCellCheckbox: SFC<
+  export const EuiTableRowCellCheckbox: FunctionComponent<
     CommonProps &
       TdHTMLAttributes<HTMLTableCellElement> &
       EuiTableRowCellCheckboxProps